Add instantiations of focal operations
