+++ source/framework/algorithm/test/slope_test.cpp
@@ -1,6 +1,6 @@
 #define BOOST_TEST_MODULE lue framework algorithm slope
 #include "lue/framework/algorithm/create_partitioned_array.hpp"
-#include "lue/framework/algorithm/focal.hpp"
+#include "lue/framework/algorithm/definition/slope.hpp"
 #include "lue/framework/test/array.hpp"
 #include "lue/framework/test/compare.hpp"
 #include "lue/framework/test/hpx_unit_test.hpp"
